drf/education/requirements.txt
django==3.2.16
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-nested-routers==0.93.4
djoser==2.1.0
djangorestframework-simplejwt
django_filter
drf_spectacular
